great_britain_system
In this code, the great Britain test system was implemented. Here, 3 functions were created which represent the 3 options of the nadir occurrence interval. The optimal set points are the ones that provide the lowest value of objective function. This code has an issue. It is: the loss of power infeed is smaller than the largest generation output.