Update 6_09_2025
In this update we have the code Scenar I_interval_3_fix_uc.  The third interval seems to be working well. The data were modified understanding why when we have a fix UC status and we make rg to be produce the most, the problem was infeasible. The reason was that the ramp limits of the nuclear power plants were zero and this didn't allow the generators to change their power set points.