back_up_6_11_2025
In this version, I figured out why RG was not procured the most. The reason for this is that you still need power to allow the electrolizers to produce hydrogen the most, reducing the headroom for generators to provide FCR. Adittionally, I modified the gaps of the codes Scenario I and Scenario I interval 3 in such a way that the results were consistent. Plots of the total procured reserve and operational setpoints of all batteries were modified and added.